Port most map logic to store
